moved the method to populate weights to be defined in brainmage model rather than pt_3dresunet
